MLFlow + Ray Serve: Display selective output for rendering
Some of the outputs, e.g. model signatures, [don't get rendered
currently](https://huggingface.co/learn/cookbook/en/mlflow_ray_serve). Presumably
this is due to them currently not being stdout text outputs.

This PR passes some of the outputs through `print()` so that they will render in
the "final" web-rendered page.

For cleanliness and explicitness, we also pass this notebook through
[nbstripout](https://github.com/kynan/nbstripout), which strips out all
outputs **except** those from cells tagged with `keep_output`. This could be
useful to apply across the board for all notebooks, if that's something that the
maintainers would be interested in.