FIX Error in VLM fine-tuning notebook
In the notebook, get_peft_model(model) is called, which mutates the base
model. Then the same model is passed to SFTTrainer with a PEFT config,
which leads to another call to get_peft_model under the hood and thus to
PEFT being applied twice.

This fix removes the manual get_peft_model call and lets trl handle
this. We lose the ability to inspect the number of trainable parameters
but IMO this is the simplest solution.